fix(net): split the TCP relay out, and re-pin the claim-13 mutation surface
Two gates CI runs that I had not run locally, both fallout from the cutover.

## check-file-size

`supervisor/flowmux.rs` reached 1557 production lines against a 1500 cap. The
per-stream TCP relay moves to `flowmux/tcp_relay.rs`, which is the same split
`udp_relay` already had and the same shape: one admitted flow, one thread that
owns the upstream socket, and the window logic that makes an exhausted send
window backpressure rather than a reset. Pure code move; `write_frame_to` and
`parse_host_port` stay in the parent because the HTTP and UDP relays share them.

The doc comment for `connect_first_admitted` travelled with the function, and
its claim that this is "the happy-eyeballs fallover used by the raw egress path"
is dropped -- that path was deleted two commits ago.

## check-mutation-witnesses

Deleting `substitution_client::substitute` took
`crates/mvm-agentd/src/substitution_client.rs` out of claim 13's mutation
surface, so the pinned baseline no longer resolved there.

Re-pinned, and worth stating why that is safe rather than routine. ADR-001's
claim-13 row names `fn:substitute`, which still resolves -- to
`keyholder/substitution.rs`, the host-side credential injection, which is where
the raw secret actually is. The function I deleted was the guest-side relay: it
framed a request carrying an opaque placeholder and never held a credential.
Its place in the surface was name matching, not coverage. Claim 13 keeps five
surface files including `keyholder/substitution.rs` and
`network_endpoint_proxy.rs`, and the re-pin removes exactly one entry.

`check-claim-catalog`, `check-witness-citations` and `check-conformance` all
still pass, which is the independent check that the ledger's named witness
survived.

Verified: 11991 workspace tests, clippy `-D warnings`, nightly fmt, the Linux
`--all-targets --all-features` cross-check, and all 57 xtask gates CI runs.
